refactor(dogstatsd): migrate serverDebug component to v2
