+++ comp/dogstatsd/bundle.go
@@ -9,7 +9,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/dogstatsd/pidmap/pidmapimpl"
 	replayfx "github.com/DataDog/datadog-agent/comp/dogstatsd/replay/fx"
 	"github.com/DataDog/datadog-agent/comp/dogstatsd/server"
-	"github.com/DataDog/datadog-agent/comp/dogstatsd/serverDebug/serverdebugimpl"
+	serverdebugfx "github.com/DataDog/datadog-agent/comp/dogstatsd/serverDebug/fx"
 	statusfx "github.com/DataDog/datadog-agent/comp/dogstatsd/status/fx"
 	offlinereporterfx "github.com/DataDog/datadog-agent/comp/offlinereporter/fx"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
@@ -21,7 +21,7 @@ import (
 func Bundle(params server.Params) fxutil.BundleOptions {
 	return fxutil.Bundle(
 		offlinereporterfx.Module(),
-		serverdebugimpl.Module(),
+		serverdebugfx.Module(),
 		replayfx.Module(),
 		pidmapimpl.Module(),
 		server.Module(params),
