--- comp/dogstatsd/serverDebug/component_mock.go
@@ -1,15 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-// Package serverdebug implements a component to run the dogstatsd server debug
-
-//go:build test
-
-package serverdebug
-
-// Mock implements mock-specific methods.
-type Mock interface {
-	Component
-}
