+++ comp/dogstatsd/serverDebug/impl/debug_test.go
@@ -19,7 +19,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	log "github.com/DataDog/datadog-agent/comp/core/log/def"
 	logmock "github.com/DataDog/datadog-agent/comp/core/log/mock"
-	serverdebug "github.com/DataDog/datadog-agent/comp/dogstatsd/serverDebug"
+	serverdebug "github.com/DataDog/datadog-agent/comp/dogstatsd/serverDebug/def"
 	"github.com/DataDog/datadog-agent/pkg/aggregator/ckey"
 	"github.com/DataDog/datadog-agent/pkg/metrics"
 	"github.com/DataDog/datadog-agent/pkg/tagset"
@@ -32,7 +32,9 @@ func fulfillDeps(t testing.TB, overrides map[string]interface{}) serverdebug.Com
 		fx.Provide(func() config.Component {
 			return config.NewMockWithOverrides(t, overrides)
 		}),
-		Module(),
+		fx.Provide(func(deps Requires) serverdebug.Component {
+			return NewComponent(deps).Comp
+		}),
 	))
 }
